update documention, remove include from headers
