+++ include/warthog/domain/gridmap.h
@@ -19,8 +19,8 @@
 //
 
 #include "grid.h"
+#include <warthog/io/fwd.h>
 #include <warthog/constants.h>
-#include <warthog/io/bittable_serialize.h>
 #include <warthog/memory/bittable.h>
 #include <warthog/util/cast.h>
 #include <warthog/util/gm_header.h>
