+++ include/warthog/io/stream_observer.h
@@ -15,7 +15,6 @@
 ///
 /// @author: Ryan Hechenberger
 /// @created: 2025-08-01
-//
 
 #include "fwd.h"
